move mock types to separate file
